+++ lib/generators/evnt/templates/initializer/app/events/application_event.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+# ApplicationEvent.
+class ApplicationEvent < Evnt::Event
+end